src/requirements.txt
debugpy==1.8.1
Django==4.2.9
djangorestframework==3.14.0
drf-nested-routers==0.93.5 #Django Nested Router
drf-spectacular==0.28.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv
